data passing to all page done
