+++ src/router.js
@@ -2,7 +2,6 @@ import React from 'react';
 import firebase from 'firebase/app';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import RankingMock from './screens/RankingMock';
 import Ranking from './screens/Ranking';
 import Info from './screens/Info';
 import SignIn from './screens/SignIn';
@@ -51,8 +50,8 @@ const Routes = () => (
         }}
       />
       <Stack.Screen
-        name="RankingMock"
-        component={RankingMock}
+        name="Ranking"
+        component={Ranking}
         options={{
           headerShown: false,
         }}
@@ -60,13 +59,6 @@ const Routes = () => (
       <Stack.Screen
         name="Home"
         component={Home}
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Stack.Screen
-        name="Ranking"
-        component={Ranking}
         options={{
           headerShown: false,
         }}
